Make SVNUrl class handle file:// URL's better.
I am not sure that this is the definitive patch, but it makes it work better.

It works most accurately with URL's like this:

file:///repos/proj/file.txt
file://d/repos/proj/file.txt

With URL's like this, "repos" is treated like the hostname instead of a segment.

file://repos/proj/file.txt
Issue #: 222