Patch from sinan@tigris.org
Adds a new preference "Fetch affected paths on demand" that is on by default.
This changes the default behavior of the Show in Resource History action to
control whether or not it retrieves the changed paths as part of the svn log
request.  In other words, the current behavior is to run the command:

svn log -v

With this patch, we would be running:

svn log

The changed paths are retrieved dynamically when you view a particular revision.

The primary goal of this patch is to reduce the memory footprint of this action
when there are a lot of changed paths in one or more of the revisions.  However,
it also should make the action run a bit faster, especially if run over a slow
network link.  The dynamic loading of the changed paths also happens quickly,
so it is actually pretty hard to notice a difference in behavior.


Issue #: 264