Updates for CL Adapter for 1.2
Implement new interface features.

Add Lock/Unlock method

Added new classes for a Notify framework.  It was using JavaHL classes but those
are not deprecated.  Removed all JavaHL dependencies.

There are some things not working.  It needs to be looked at some more.  I do not
use this a lot so I do not know if all of these things were working before.