Show Annotations was not working if the file had not been modified yet in the current location.  This would most typically occur when doing it from a branch.
In JavaHL we had to to switch to using the new syntax that allows us to
specify a peg revision.  In this case, we specify a peg revision of HEAD
which lets Subversion trace the history back and find where the file came from.

In the CLI, it didn't seem to be working at all because it was specifying a
revision range of -r START:rev_num and START is not a valid revision.  I fixed
this by changing it to just specify -r rev_num which is the same thing.  The
CLI automatically specifies a peg revision of @HEAD so there was no need to
specifically supply it.