Fix a potential NPE in SVNNotificationHandler when using JavaSVN.
This appears to be a bug in JavaSVN but a commit fails and it doesn't throw
an Exception with the error, but it does send a notification with an action
of -1 and a path of null.  The null causes an NPE in the code that sends
file notifications.

I added a check for the -1 and send an error message that the command ended
abnormally and I also added the NPE checks in as well.