Remove the SASL auth mechanisms.
I was hoping if we shipped and pre-loaded them that SASL would find them.  It
looks like it still needs to registry key to be set and point at their location.
So the best option is to have people set this up outside Subclipse, by installing
something like CollabNet Subversion.  The fact that we include a SASL-aware
version of JavaHL will still make this a lot easier.